requirements.txt
# Core Framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.0.0,<3.0.0
pydantic-settings>=2.0.0,<3.0.0

# LLM Clients
groq
google-generativeai

# Async Support
asyncio
aiofiles

# Cache
redis

# MongoDB (Optional - for storing council outputs)
# Note: motor includes pymongo as a dependency, so we only need motor
motor>=3.3.0

# Environment Variables
python-dotenv

# HTTP Clients
httpx
requests

# Data Validation & Serialization
pydantic[email]

# Utilities
python-multipart

# Testing (Optional)
pytest
pytest-asyncio
pytest-cov

# Development Tools (Optional)
black
flake8
mypy
isort

# Logging & Monitoring
colorlog

# Production Server (Optional)
gunicorn